\name{bmps_decide}
\alias{bmps_decide}
\title{Turn state calls into progenitor calls}
\description{Turn state calls into progenitor calls.}
\usage{bmps_decide(calls, min_margin = 100, min_top1 = 0)}
\arguments{
  \item{calls}{Data frame from \code{bmps_calls}.}
  \item{min_margin}{Minimum margin.}
  \item{min_top1}{Minimum top1.}
}
\value{A character vector of calls.}
